+++ src/main/java/com/example/fastcampusmysql/domain/member/repository/JpaMemberRepository.java
@@ -1,0 +1,7 @@
+package com.example.fastcampusmysql.domain.member.repository;
+
+import com.example.fastcampusmysql.domain.member.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface JpaMemberRepository extends JpaRepository<Member, Long>, MemberRepository {
+}
