+++ src/main/java/com/example/fastcampusmysql/domain/member/repository/MemberNicknameHistoryRepository.java
@@ -1,61 +1,10 @@
 package com.example.fastcampusmysql.domain.member.repository;
 
 import com.example.fastcampusmysql.domain.member.entity.MemberNicknameHistory;
-import lombok.RequiredArgsConstructor;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
-import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
-import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
-import org.springframework.jdbc.core.namedparam.SqlParameterSource;
-import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import org.springframework.stereotype.Repository;
 
-import java.sql.ResultSet;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.List;
-import java.util.Optional;
 
-@RequiredArgsConstructor
-@Repository
-public class MemberNicknameHistoryRepository {
-    private final NamedParameterJdbcTemplate namedParameterJdbcTemplate;
-    private static final String TABLE = "MemberNicknameHistory";
-    private static final RowMapper<MemberNicknameHistory> rowMapper = (ResultSet resultSet, int rowNum) -> MemberNicknameHistory
-            .builder()
-            .id(resultSet.getLong("id"))
-            .memberId(resultSet.getLong("memberId"))
-            .nickname(resultSet.getString("nickname"))
-            .createdAt(resultSet.getObject("createdAt", LocalDateTime.class))
-            .build();
-
-    public List<MemberNicknameHistory> findAllByMemberId(Long memberId) {
-        String sql = String.format("SELECT * FROM %s WHERE memberId = :memberId", TABLE);
-        SqlParameterSource params = new MapSqlParameterSource().addValue("memberId", memberId);
-        return namedParameterJdbcTemplate.query(sql, params, rowMapper);
-    }
-    public MemberNicknameHistory save(MemberNicknameHistory memberNicknameHistory) {
-        /*
-            member id 보고 갱신 또는 삽입을 정함
-            반환값은 id를 담아서 반환한다.
-         */
-        if (memberNicknameHistory.getId() == null) {
-            return insert(memberNicknameHistory);
-        }
-        throw new UnsupportedOperationException("MemberNicknameHistory는 갱신을 지원하지 않습니다");
-    }
-    private MemberNicknameHistory insert(MemberNicknameHistory memberNicknameHistory) {
-        SimpleJdbcInsert simpleJdbcInsert = new SimpleJdbcInsert(namedParameterJdbcTemplate.getJdbcTemplate())
-                .withTableName("MemberNicknameHistory")
-                .usingGeneratedKeyColumns("id");
-        SqlParameterSource params = new BeanPropertySqlParameterSource(memberNicknameHistory);
-        Long id = simpleJdbcInsert.executeAndReturnKey(params).longValue();
-        return MemberNicknameHistory
-                .builder()
-                .id(id)
-                .memberId(memberNicknameHistory.getMemberId())
-                .nickname(memberNicknameHistory.getNickname())
-                .createdAt(memberNicknameHistory.getCreatedAt())
-                .build();
-    }
+public interface MemberNicknameHistoryRepository {
+    List<MemberNicknameHistory> findAllByMemberId(Long memberId);
+    MemberNicknameHistory save(MemberNicknameHistory memberNicknameHistory);
 }
