+++ src/main/java/com/example/fastcampusmysql/domain/member/repository/MemberRepository.java
@@ -1,80 +1,12 @@
 package com.example.fastcampusmysql.domain.member.repository;
 
 import com.example.fastcampusmysql.domain.member.entity.Member;
-import lombok.RequiredArgsConstructor;
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.jdbc.core.namedparam.BeanPropertySqlParameterSource;
-import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
-import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
-import org.springframework.jdbc.core.namedparam.SqlParameterSource;
-import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
-import org.springframework.stereotype.Repository;
 
-import java.sql.ResultSet;
-import java.time.LocalDate;
-import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
-@RequiredArgsConstructor
-@Repository
-public class MemberRepository {
-    final private NamedParameterJdbcTemplate namedParameterJdbcTemplate;
-    static final private String TABLE = "Member";
-    private static final RowMapper<Member> rowMapper = (ResultSet resultSet, int rowNum) -> Member
-            .builder()
-            .id(resultSet.getLong("id"))
-            .email(resultSet.getString("email"))
-            .nickname(resultSet.getString("nickname"))
-            .birthday(resultSet.getObject("birthday", LocalDate.class))
-            .createdAt(resultSet.getObject("createdAt", LocalDateTime.class))
-            .build();
-    public Optional<Member> findById(Long id) {
-        String sql = String.format("select * from %s WHERE id = :id", TABLE);
-        SqlParameterSource param = new MapSqlParameterSource()
-                .addValue("id", id);
-        Member member = namedParameterJdbcTemplate.queryForObject(sql, param, rowMapper);
-        return Optional.ofNullable(member);
-    }
-
-    public List<Member> findAllbyIdIn(List<Long> ids) {
-        if (ids.isEmpty()) {
-            return List.of();
-        }
-        String sql = String.format("SELECT * FROM %s WHERE id in (:ids)", TABLE);
-        SqlParameterSource params = new MapSqlParameterSource().addValue("ids", ids);
-        return namedParameterJdbcTemplate.query(sql, params, rowMapper);
-    }
-    public Member save(Member member) {
-        /*
-            member id 보고 갱신 또는 삽입을 정함
-            반환값은 id를 담아서 반환한다.
-         */
-        if (member.getId() == null) {
-            return insert(member);
-        }
-        return update(member);
-    }
-    private Member insert(Member member) {
-        SimpleJdbcInsert simpleJdbcInsert = new SimpleJdbcInsert(namedParameterJdbcTemplate.getJdbcTemplate())
-                .withTableName("Member")
-                .usingGeneratedKeyColumns("id");
-        SqlParameterSource params = new BeanPropertySqlParameterSource(member);
-        Long id = simpleJdbcInsert.executeAndReturnKey(params).longValue();
-        return Member
-                .builder()
-                .id(id)
-                .nickname(member.getNickname())
-                .email(member.getEmail())
-                .birthday(member.getBirthday())
-                .createdAt(member.getCreatedAt())
-                .build();
-    }
-    private Member update(Member member) {
-        // TODO: sql 반복 되는 것을 제거해보기
-        String sql = String.format("UPDATE %s SET email = :email, nickname = :nickname, birthday = :birthday WHERE id = :id", TABLE);
-        SqlParameterSource params = new BeanPropertySqlParameterSource(member);
-        namedParameterJdbcTemplate.update(sql, params);
-        return member;
-    }
+public interface MemberRepository {
+    Optional<Member> findById(Long id);
+    List<Member> findAllByIdIn(List<Long> ids);
+    Member save(Member member);
 }
