+++ src/main/java/com/example/fastcampusmysql/domain/member/service/MemberWriteService.java
@@ -7,12 +7,14 @@ import com.example.fastcampusmysql.domain.member.repository.MemberNicknameHistor
 import com.example.fastcampusmysql.domain.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @RequiredArgsConstructor
 @Service
 public class MemberWriteService {
     final private MemberRepository memberRepository;
     private final MemberNicknameHistoryRepository memberNicknameHistoryRepository;
+    @Transactional
     public Member register(RegisterMemberCommand registerMemberCommand) {
         /*
             목표 - 회원정보(이메일, 닉네임, 생년월일)를 등록한다.
@@ -32,6 +34,7 @@ public class MemberWriteService {
         return saveMember;
     }
 
+    @Transactional
     public void changeNickname(Long memberId, String nickname) {
         Member member = memberRepository.findById(memberId).orElseThrow();
         member.changeNickname(nickname);
